Merge lens in poller
This removes the need for the source field which was a work around for
not being able to merge values.

Additionally the elapsed used to normalize the value is no longer
accurate as there's no real way to remain accurate while merge the
values of multiple optics instance with different elapsed time. This
might introduce slight glitch in the graph when the poller is restarted
which is considered acceptable for now.